--- src/features/dashboard/super/admin-permissions/organisms/index.ts
@@ -1,1 +1,0 @@
-export * from "./PermissionList";
