--- src/features/dashboard/super/admin-permissions/templates/index.ts
@@ -1,1 +1,0 @@
-export * from "./AdminPermissions";
