+++ src/features/dashboard/super/configure-permissions/templates/ConfigurePermissions.db.ts
