+++ src/features/dashboard/super/configure-permissions/templates/index.ts
@@ -1,0 +1,1 @@
+export * from "./ConfigurePermissions";
