+++ src/features/dashboard/super/grant-permissions/organisms/index.ts
@@ -1,1 +1,1 @@
-export * from "./AdminCard";
+export * from "./Admin";
